requirements.txt
alembic==1.7.5
aniso8601==9.0.1
astroid==2.11.7
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.0
black==22.6.0
cffi==1.15.0
click==7.1.2
cryptography==36.0.1
dill==0.3.5.1
Flask==1.1.4
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Migrate==2.1.1
flask-restplus==0.13.0
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
importlib-metadata==4.10.1
importlib-resources==5.4.0
isort==5.10.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.4.0
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
PyJWT==1.7.1
pylama==8.3.8
pylint==2.14.5
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
pyOpenSSL==22.0.0
pyrsistent==0.18.1
pytz==2021.3
six==1.16.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.31
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
Werkzeug==1.0.1
wrapt==1.14.1
yapf==0.32.0
zipp==3.7.0
